+++ src/EFCore.Ydb/src/Properties/AssemblyInfo.cs
@@ -1,3 +1,5 @@
+using System.Runtime.CompilerServices;
 using Microsoft.EntityFrameworkCore.Design;
 
 [assembly: DesignTimeProviderServices("EntityFrameworkCore.Ydb.Design.Internal.YdbDesignTimeServices")]
+[assembly: InternalsVisibleTo("EntityFrameworkCore.Ydb.FunctionalTests")]