+++ src/Ydb.Sdk/test/Ydb.Sdk.Ado.Tests/Utils/MockAsyncEnumerator.cs
@@ -1,6 +1,6 @@
 namespace Ydb.Sdk.Ado.Tests.Utils;
 
-public class MockAsyncEnumerator<T>(IEnumerable<T> items) : IServerStream<T>
+internal class MockAsyncEnumerator<T>(IEnumerable<T> items) : IServerStream<T>
 {
     private readonly IEnumerator<T> _inner = items.GetEnumerator();