+++ src/Ydb.Sdk/test/Ydb.Sdk.Topic.Tests/IDriverFactoryMock.cs
@@ -4,7 +4,7 @@ using Ydb.Sdk.Ado;
 
 namespace Ydb.Sdk.Topic.Tests;
 
-public class IDriverFactoryMock(Mock<IDriver> mockIDriver, string grpcConnectionString) : IDriverFactory
+internal class IDriverFactoryMock(Mock<IDriver> mockIDriver, string grpcConnectionString) : IDriverFactory
 {
     public Task<IDriver> CreateAsync() => Task.FromResult(mockIDriver.Object);